requirements.txt
Flask
Flask-JWT-Extended
Flask-WTF
Flask-MySQLdb
bcrypt
python-dotenv
gunicorn